% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abraham2024}
\alias{abraham2024}
\alias{Abraham2024}
\title{Abraham et al. 2024
Abraham et al. (2024) (\doi{10.1016/j.envint.2024.109047}) determined the 
half-lives of 15 per- and 
polyfluoroalkyl substances in a single male volunteer.}
\format{
data.frame
}
\source{
Wambaugh et al., Applying High Throughput Toxicokinetics (HTTK) to Per- and
Polyfluoro Alkyl Substances (PFAS), submitted
}
\usage{
abraham2024
}
\description{
Abraham et al. 2024
Abraham et al. (2024) (\doi{10.1016/j.envint.2024.109047}) determined the 
half-lives of 15 per- and 
polyfluoroalkyl substances in a single male volunteer.
}
\references{
\insertRef{abraham2024kinetics}{httkexamples}
}
\keyword{data}
