% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{armitage_input}
\alias{armitage_input}
\alias{Armitage_input}
\title{Parameters for in vitro distribution analysis in Honda et al. (2019)
Honda et al. (2019) (\doi{10.1371/journal.pone.0217564}) used the
Armitage et al. (2014) (\doi{10.1021/es501955g}) mass-balance model to 
predict the impact of
in vitro partitioning on free chemical concentrations.}
\format{
data.frame
}
\source{
\insertRef{honda2019using}{httkexamples}
}
\usage{
armitage_input
}
\description{
Parameters for in vitro distribution analysis in Honda et al. (2019)
Honda et al. (2019) (\doi{10.1371/journal.pone.0217564}) used the
Armitage et al. (2014) (\doi{10.1021/es501955g}) mass-balance model to 
predict the impact of
in vitro partitioning on free chemical concentrations.
}
\references{
\insertRef{armitage2014application}{httkexamples}
\insertRef{honda2019using}{httkexamples}
}
\keyword{data}
