% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dermal.nonvolatilechems}
\alias{dermal.nonvolatilechems}
\title{Non-volatile chemicals with ToxCast data
Meade et al. (submitted) performed in vitro-in vivo extrapolation for 
dermal exposures assuming 8 hours of exposure via hands submerged in a
liquid with 1 ppm of chemical. These were the chemicals analyzed.}
\format{
data.frame

data.frame
}
\source{
Meade et al., Incorporating a dermal absorption
route into high throughput toxicokinetic modeling, submitted.

Meade et al., Incorporating a dermal absorption
route into high throughput toxicokinetic modeling, submitted.
}
\usage{
dermal.nonvolatilechems

dermal.nonvolatilechems
}
\description{
Non-volatile chemicals with ToxCast data
Meade et al. (submitted) performed in vitro-in vivo extrapolation for 
dermal exposures assuming 8 hours of exposure via hands submerged in a
liquid with 1 ppm of chemical. These were the chemicals analyzed.

Non-volatile chemicals with ToxCast data
Meade et al. (submitted) performed in vitro-in vivo extrapolation for 
dermal exposures assuming 8 hours of exposure via hands submerged in a
liquid with 1 ppm of chemical. These were the chemicals analyzed.
}
\references{
Meade et al., Incorporating a dermal absorption
route into high throughput toxicokinetic modeling, submitted.

Meade et al., Incorporating a dermal absorption
route into high throughput toxicokinetic modeling, submitted.
}
\keyword{data}
