% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dermal.toxcast}
\alias{dermal.toxcast}
\title{Chemicals with ToxCast data for Meade et al. (submitted) chemicals
Meade et al. (submitted) performed in vitro-in vivo extrapolation for 
dermal exposures assuming 8 hours of exposure via hands submerged in a
liquid with 1 ppm of chemical. These are the ToxCast in vitro screening
data for those chemicals.}
\format{
data.frame
}
\source{
https://www.epa.gov/comptox-tools/exploring-toxcast-data
}
\usage{
dermal.toxcast
}
\description{
Chemicals with ToxCast data for Meade et al. (submitted) chemicals
Meade et al. (submitted) performed in vitro-in vivo extrapolation for 
dermal exposures assuming 8 hours of exposure via hands submerged in a
liquid with 1 ppm of chemical. These are the ToxCast in vitro screening
data for those chemicals.
}
\references{
Meade et al., Incorporating a dermal absorption
route into high throughput toxicokinetic modeling, submitted.
}
\keyword{data}
