% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dermalCvT2025}
\alias{dermalCvT2025}
\title{Toxicokinetic concentration vs. time (CvT) data for Meade et al. (submitted) chemicals
Meade et al. (submitted) evaluated a generic PBTK model for dermal exposure
using in vivo CvT data curated from the literature. These data will
eventually be incorporated in the the CvTdb (Sayre et al., 2020, \doi{10.1038/s41597-020-0455-1}).}
\format{
data.frame
}
\source{
\insertRef{sayre2020database}{httkexamples}
}
\usage{
dermalCvT2025
}
\description{
Toxicokinetic concentration vs. time (CvT) data for Meade et al. (submitted) chemicals
Meade et al. (submitted) evaluated a generic PBTK model for dermal exposure
using in vivo CvT data curated from the literature. These data will
eventually be incorporated in the the CvTdb (Sayre et al., 2020, \doi{10.1038/s41597-020-0455-1}).
}
\references{
Meade et al., Incorporating a dermal absorption
route into high throughput toxicokinetic modeling, submitted.
\insertRef{sayre2020database}{httkexamples}
}
\keyword{data}
