% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{howgate}
\alias{howgate}
\title{Howgate et al. (2006)}
\format{
A data.table containing 24 rows and 11 columns.

A data.table containing 24 rows and 11 columns.
}
\usage{
howgate

howgate
}
\description{
This data set is only used in Vignette 5.

This data set is only used in Vignette 5.
}
\references{
Howgate, E. M., et al. "Prediction of in vivo drug clearance from in vitro
data. I: impact of inter-individual variability." Xenobiotica 36.6 (2006):
473-497.

Howgate, E. M., et al. "Prediction of in vivo drug clearance from in vitro
data. I: impact of inter-individual variability." Xenobiotica 36.6 (2006):
473-497.
}
\author{
Caroline Ring
}
\keyword{data}
