% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{huh2011}
\alias{huh2011}
\alias{Huh2011}
\title{Huh et al. 2011
Huh et al. (2011) (\doi{10.3109/00498254.2011.598582}) provided interspecies
allometric scaling parameters for whole body clearance for a a variety of 
pharmaceuticals.}
\format{
data.frame
}
\source{
Wambaugh et al., Applying High Throughput Toxicokinetics (HTTK) to Per- and
Polyfluoro Alkyl Substances (PFAS), submitted
}
\usage{
huh2011
}
\description{
Huh et al. 2011
Huh et al. (2011) (\doi{10.3109/00498254.2011.598582}) provided interspecies
allometric scaling parameters for whole body clearance for a a variety of 
pharmaceuticals.
}
\references{
\insertRef{huh2011interspecies}{httkexamples}
}
\keyword{data}
