% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{johnson}
\alias{johnson}
\title{Johnson et al. (2006)}
\format{
A data.table containing 60 rows and 11 columns.

A data.table containing 60 rows and 11 columns.
}
\usage{
johnson

johnson
}
\description{
This data set is only used in Vignette 5.

This data set is only used in Vignette 5.
}
\references{
Johnson, Trevor N., Amin Rostami-Hodjegan, and Geoffrey T. Tucker.
"Prediction of the clearance of eleven drugs and associated variability in
neonates, infants and children." Clinical pharmacokinetics 45.9 (2006):
931-956.

Johnson, Trevor N., Amin Rostami-Hodjegan, and Geoffrey T. Tucker.
"Prediction of the clearance of eleven drugs and associated variability in
neonates, infants and children." Clinical pharmacokinetics 45.9 (2006):
931-956.
}
\author{
Caroline Ring
}
\keyword{data}
