% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meade2025}
\alias{meade2025}
\alias{Meade2025}
\title{Simulation outputs from Meade et al. (submitted)
Meade et al. (submitted) performed generic PBTK simulations for dermal
exposure under a variety of assumptions. Although the code to recreate these
simulations is provided, it is time-intensive. The 2025 outputs from the 
simulations are stored in this list of data.frames.}
\format{
list
}
\source{
Meade et al., Incorporating a dermal absorption
route into high throughput toxicokinetic modeling, submitted.
}
\usage{
meade2025
}
\description{
Simulation outputs from Meade et al. (submitted)
Meade et al. (submitted) performed generic PBTK simulations for dermal
exposure under a variety of assumptions. Although the code to recreate these
simulations is provided, it is time-intensive. The 2025 outputs from the 
simulations are stored in this list of data.frames.
}
\references{
Meade et al., Incorporating a dermal absorption
route into high throughput toxicokinetic modeling, submitted.
}
\keyword{data}
