% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metabolism_data_Linakis2020}
\alias{metabolism_data_Linakis2020}
\title{Metabolism data involved in Linakis et al. 2020 (\doi{10.1038/s41370-020-0238-y}) vignette analysis.}
\format{
A data.frame containing x rows and y columns.

A data.frame containing x rows and y columns.
}
\source{
Matt Linakis

Matt Linakis
}
\usage{
metabolism_data_Linakis2020

metabolism_data_Linakis2020
}
\description{
Metabolism data involved in Linakis et al. 2020 (\doi{10.1038/s41370-020-0238-y}) vignette analysis.

Metabolism data involved in Linakis 2020 vignette analysis.
}
\references{
\insertRef{linakis2020development}{httkexamples}

\insertRef{linakis2020development}{httkexamples}
}
\author{
Matt Linakis
}
\keyword{data}
