% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pc.data}
\alias{pc.data}
\title{Partition Coefficient Data}
\format{
A data.frame.

A data.frame.
}
\usage{
pc.data

pc.data
}
\description{
Measured rat in vivo partition coefficients and data for predicting them.

Measured rat in vivo partition coefficients and data for predicting them.
}
\references{
\insertRef{schmitt2008general}{httkexamples}

\insertRef{schmitt2008corrigendum}{httkexamples}

Poulin, P. and F.P. Theil, A priori prediction of tissue: plasma partition
coefficients of drugs to facilitate the use of physiologically based
pharmacokinetic models in drug discovery. Journal of pharmaceutical
sciences, 2000. 89(1): p. 16-35.

Rodgers, T. and M. Rowland, Physiologically based pharmacokinetic modelling
2: predicting the tissue distribution of acids, very weak bases, neutrals
and zwitterions. Journal of pharmaceutical sciences, 2006. 95(6): p.
1238-1257.

Rodgers, T., D. Leahy, and M. Rowland, Physiologically based pharmacokinetic
modeling 1: predicting the tissue distribution of moderate-to-strong bases.
Journal of pharmaceutical sciences, 2005. 94(6): p. 1259-1276.

Rodgers, T., D. Leahy, and M. Rowland, Tissue distribution of basic drugs:
Accounting for enantiomeric, compound and regional differences amongst
beta-blocking drugs in rat. Journal of pharmaceutical sciences, 2005. 94(6):
p. 1237-1248.

Gueorguieva, I., et al., Development of a whole body physiologically based
model to characterise the pharmacokinetics of benzodiazepines. 1: Estimation
of rat tissue-plasma partition ratios. Journal of pharmacokinetics and
pharmacodynamics, 2004. 31(4): p. 269-298.

Poulin, P., K. Schoenlein, and F.P. Theil, Prediction of adipose tissue:
plasma partition coefficients for structurally unrelated drugs. Journal of
pharmaceutical sciences, 2001. 90(4): p. 436-447.

Bjorkman, S., Prediction of the volume of distribution of a drug: which
tissue-plasma partition coefficients are needed? Journal of pharmacy and
pharmacology, 2002. 54(9): p. 1237-1245.

\insertRef{yun2013correlation}{httkexamples}

Uchimura, T., et al., Prediction of human blood-to-plasma drug concentration
ratio. Biopharmaceutics & drug disposition, 2010. 31(5-6): p. 286-297.

\insertRef{schmitt2008general}{httkexamples}

\insertRef{schmitt2008corrigendum}{httkexamples}

Poulin, P. and F.P. Theil, A priori prediction of tissue: plasma partition
coefficients of drugs to facilitate the use of physiologically based
pharmacokinetic models in drug discovery. Journal of pharmaceutical
sciences, 2000. 89(1): p. 16-35.

Rodgers, T. and M. Rowland, Physiologically based pharmacokinetic modelling
2: predicting the tissue distribution of acids, very weak bases, neutrals
and zwitterions. Journal of pharmaceutical sciences, 2006. 95(6): p.
1238-1257.

Rodgers, T., D. Leahy, and M. Rowland, Physiologically based pharmacokinetic
modeling 1: predicting the tissue distribution of moderate-to-strong bases.
Journal of pharmaceutical sciences, 2005. 94(6): p. 1259-1276.

Rodgers, T., D. Leahy, and M. Rowland, Tissue distribution of basic drugs:
Accounting for enantiomeric, compound and regional differences amongst
beta-blocking drugs in rat. Journal of pharmaceutical sciences, 2005. 94(6):
p. 1237-1248.

Gueorguieva, I., et al., Development of a whole body physiologically based
model to characterise the pharmacokinetics of benzodiazepines. 1: Estimation
of rat tissue-plasma partition ratios. Journal of pharmacokinetics and
pharmacodynamics, 2004. 31(4): p. 269-298.

Poulin, P., K. Schoenlein, and F.P. Theil, Prediction of adipose tissue:
plasma partition coefficients for structurally unrelated drugs. Journal of
pharmaceutical sciences, 2001. 90(4): p. 436-447.

Bjorkman, S., Prediction of the volume of distribution of a drug: which
tissue-plasma partition coefficients are needed? Journal of pharmacy and
pharmacology, 2002. 54(9): p. 1237-1245.

\insertRef{yun2013correlation}{httkexamples}

Uchimura, T., et al., Prediction of human blood-to-plasma drug concentration
ratio. Biopharmaceutics & drug disposition, 2010. 31(5-6): p. 286-297.
}
\author{
Jimena Davis and Robert Pearce
}
\keyword{data}
