% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pksim.pcs}
\alias{pksim.pcs}
\title{Partition Coefficients from PK-Sim}
\format{
data.frame

data.frame
}
\source{
\insertRef{kapraun2022fetalmodel}{httkexamples}

\insertRef{kapraun2022fetalmodel}{httkexamples}
}
\usage{
pksim.pcs

pksim.pcs
}
\description{
Dallmann et al. (2018) made use of PK-Sim to predict chemical- and tissue-
specific partition coefficients. The methods include both the default
PK-Sim approach and PK-Sim Standard and Rodgers & Rowland (2006).

Dallmann et al. (2018) made use of PK-Sim to predict chemical- and tissue-
specific partition coefficients. The methods include both the default
PK-Sim approach and PK-Sim Standard and Rodgers & Rowland (2006).
}
\references{
\insertRef{dallmann2018pregpbtk}{httkexamples}

\insertRef{dallmann2018pregpbtk}{httkexamples}
}
\keyword{data}
