% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supptab2_Linakis2020}
\alias{supptab2_Linakis2020}
\title{More supplementary output from Linakis 2020 vignette analysis.}
\format{
A data.frame containing x rows and y columns.

A data.frame containing x rows and y columns.
}
\source{
Matt Linakis

Matt Linakis
}
\usage{
supptab2_Linakis2020

supptab2_Linakis2020
}
\description{
More supplementary output from Linakis 2020 vignette analysis.

More supplementary output from Linakis 2020 vignette analysis.
}
\references{
\insertRef{linakis2020development}{httkexamples}

\insertRef{linakis2020development}{httkexamples}
}
\author{
Matt Linakis
}
\keyword{data}
