% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wallis2023}
\alias{wallis2023}
\alias{Wallis2023}
\title{Wallis et al. 2023
Wallis et al. (2023) (\doi{10.1021/acs.est.2c08241}) estimated the human
toxicokinetic half-lives for a range of per- and poly-fluorinated alkyl 
substances (PFAS).}
\format{
data.frame
}
\source{
Wambaugh et al., Applying High Throughput Toxicokinetics (HTTK) to Per- and
Polyfluoro Alkyl Substances (PFAS), submitted
}
\usage{
wallis2023
}
\description{
Wallis et al. 2023
Wallis et al. (2023) (\doi{10.1021/acs.est.2c08241}) estimated the human
toxicokinetic half-lives for a range of per- and poly-fluorinated alkyl 
substances (PFAS).
}
\references{
\insertRef{wallis2023estimation}{httkexamples}
}
\keyword{data}
