% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wambaugh2019.nhanes}
\alias{wambaugh2019.nhanes}
\title{NHANES Chemical Intake Rates for chemicals in Wambaugh et al. (2019)}
\format{
A data frame with 20 rows and 4 variables:
\describe{
  \item{lP}{The median of the Bayesian credible interval for median population
intake rate (mg/kg bodyweight/day)}
  \item{lP.min}{The lower 95th percentile of the Bayesian credible interval for median population
intake rate (mg/kg bodyweight/day)}
  \item{lP.max}{The upper 95th percentile of the Bayesian credible interval for median population
intake rate (mg/kg bodyweight/day)}
  \item{CASRN}{The Chemical Abstracts Service Registry Number}
}

A data frame with 20 rows and 4 variables:
\describe{
  \item{lP}{The median of the Bayesian credible interval for median population
intake rate (mg/kg bodyweight/day)}
  \item{lP.min}{The lower 95th percentile of the Bayesian credible interval for median population
intake rate (mg/kg bodyweight/day)}
  \item{lP.max}{The upper 95th percentile of the Bayesian credible interval for median population
intake rate (mg/kg bodyweight/day)}
  \item{CASRN}{The Chemical Abstracts Service Registry Number}
}
}
\source{
Wambaugh et al. (2019)

Wambaugh et al. (2019)
}
\usage{
wambaugh2019.nhanes

wambaugh2019.nhanes
}
\description{
These data are a subset of the Bayesian inferences reported by Ring et al.
(2017) (\doi{10.1016/j.envint.2017.06.004}) from the U.S. Centers for Disease Control and Prevention (CDC)
National Health and Nutrition Examination Survey (NHANES). They reflect the
population median intake rate (mg/kg body weight/day), with uncertainty.

These data are a subset of the Bayesian inferrences reported by Ring et al.
(2017) from the U.S. Centers for Disease Control and Prevention (CDC)
National Health and Nutrition Examination Survey (NHANES). They reflect the
populaton median intake rate (mg/kg body weight/day), with uncertainty.
}
\references{
\insertRef{ring2017identifying}{httkexamples}

\insertRef{wambaugh2019assessing}{httkexamples}

\insertRef{ring2017identifying}{httkexamples}

\insertRef{wambaugh2019assessing}{httkexamples}
}
\author{
John Wambaugh
}
\keyword{data}
