% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wang2018}
\alias{wang2018}
\alias{Wang2018}
\title{Wang et al. 2018
Wang et al. (2018) screened the blood of 75 pregnant women for the presence 
of environmental organic acids (EOAs) and identified mass spectral features 
corresponding to 453 chemical formulae of which 48 could be mapped to likely 
structures. Of the 48 with tentative structures the identity of six were 
confirmed with available chemical standards.}
\format{
data.frame

data.frame
}
\source{
\insertRef{kapraun2022fetalmodel}{httkexamples}

\insertRef{kapraun2022fetalmodel}{httkexamples}
}
\usage{
wang2018

wang2018
}
\description{
Wang et al. 2018
Wang et al. (2018) screened the blood of 75 pregnant women for the presence 
of environmental organic acids (EOAs) and identified mass spectral features 
corresponding to 453 chemical formulae of which 48 could be mapped to likely 
structures. Of the 48 with tentative structures the identity of six were 
confirmed with available chemical standards.

Wang et al. 2018
Wang et al. (2018) (\doi{10.1289/EHP2920}) screened the blood of 75 pregnant women for the presence 
of environmental organic acids (EOAs) and identified mass spectral features 
corresponding to 453 chemical formulae of which 48 could be mapped to likely 
structures. Of the 48 with tentative structures the identity of six were 
confirmed with available chemical standards.
}
\references{
\insertRef{Wang2018matbloodnta}{httkexamples}

\insertRef{Wang2018matbloodnta}{httkexamples}
}
\keyword{data}
