% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Wetmore2012}
\alias{Wetmore2012}
\title{Published toxicokinetic predictions based on in vitro data from Wetmore et
al. 2012.}
\format{
A data.frame containing 13 rows and 15 columns.

A data.frame containing 13 rows and 15 columns.
}
\usage{
Wetmore2012

Wetmore2012
}
\description{
This data set overlaps with Wetmore.data and is used only in Vignette 4 for
steady state concentration.

This data set overlaps with Wetmore.data and is used only in Vignette 4 for
steady state concentration.
}
\references{
\insertRef{wetmore2012integration}{httkexamples}

\insertRef{wetmore2012integration}{httkexamples}
}
\keyword{data}
